bugfix: show 12 o'clock correctly
previously, any time between 12:00 and 12:59 would be displayed as
0 through 59, with no digits other than the minute shown. this was
because `currentTime.hour() % 12` would for obvious reasons come
out to be 0 whenever the hour was 12 o'clock, and showNumberDecEx
would duly strip the leading zeros from the resulting number we
would try to display. fix this by manually checking to see if the
hour is 0 and setting it to 12.

an alternative way to do this without if statements would be to
change:

currentTime.hour() % 12

to:

((currentTime.hour() + 11) % 12) + 1

but that's a bit too clever for my taste, especially in a project
that's otherwise supposed to be accessible to beginners.